// src/interfaces/orderbook.ts
interface Ask {
  price: number;
  quantity: number;
}
export interface RequestOrder {
  asks: Ask[];
  bids: Ask[];
  timestamp: number;
}
// ...
export interface Holding {
  token: string;
  holding: number;
  frozen: number;
  pending_short: number;
  updated_time: number;
}
// ...
export interface Balance {
  holding: number;
  frozen: number;
  interest: number;
  pendingShortQty: number;
  pendingExposure: number;
  pendingLongQty: number;
  pendingLongExposure: number;
  version: number;
  staked: number;
  unbonding: number;
  vault: number;
  averageOpenPrice: number;
  pnl24H: number;
  fee24H: number;
  markPrice: number;
}

export interface OrderTrade {
  id: number;
  symbol: string;
  fee: number;
  fee_asset: string;
  side: string;
  order_id: number;
  executed_price: number;
  executed_quantity: number;
  executed_timestamp: number;
  is_maker: number;
}

export interface SymbolInfo {
  symbol: string;
  quote_min: number;
  quote_max: number;
  quote_tick: number;
  base_min: number;
  base_max: number;
  base_tick: number;
  min_notional: number;
  price_range: number;
  price_scope: number;
  std_liquidation_fee: number;
  liquidator_fee: number;
  claim_insurance_fund_discount: number;
  funding_period: number;
  cap_funding: number;
  floor_funding: number;
  interest_rate: number;
  created_time: number;
  updated_time: number;
  imr_factor: number;
  base_mmr: number;
  base_imr: number;
}